Fixed bug in NodeHeap
Fixed a bug in the NodeHeap class that was allowing duplicate nodes to be
pushed onto the heap, bumping out other nodes.

Sending/Retreiving messages appears to be working like it should, but more
testing is needed. Messages were not always stored in the correct neighborhood,
but I tweaked my test setup and was able to sucessfully send and retrive a
number of messages in a row.