Change k to 40
The way the protocol works is the sender needs to set the message key wiithin
the same range as the recipient's public key. The problem is the sender doesn't
know what that range is. We could do a find_node RPC to find the range but then
we a leaking info about who we are trying to communicate with.

As an alternative subspace uses the size of the range of our own neighborhood
as a proxy for the size of the range of the recipient. If nodes were evenly
distributed across the network, this would work fine every time, but that is not
the case. Sometimes the message key ends up falling outside the correct range.

To compensate for these errors we're going to have to increase the k value.
Setting it to 40 should reduce our error rate pretty close to zero.

Alternatively we could use a smaller range when setting the key, but that
come at the expense of a smaller anonymity set.