Switched to 256 bit DHT
Switch the code base over to a 256 bit DHT to allow for larger anonymity sets.
Previously there were about 2^19 possible keys in each neighborhood (depending
on the number of nodes in the network). Now it's about 2^31.

This change should also reduce the amount of brute forcing used when sending a message.
It was trivial before, but now it's even more so.